Move basic components to ui-mobile
