+++ ui-app/src/jbMain/kotlin/com/alexvanyo/composelife/ui/app/action/settings/DarkThemeConfigUi.kt
@@ -27,13 +27,13 @@ import com.alexvanyo.composelife.preferences.DarkThemeConfig
 import com.alexvanyo.composelife.preferences.di.ComposeLifePreferencesProvider
 import com.alexvanyo.composelife.preferences.di.LoadedComposeLifePreferencesProvider
 import com.alexvanyo.composelife.preferences.setDarkThemeConfig
-import com.alexvanyo.composelife.ui.app.component.DropdownOption
-import com.alexvanyo.composelife.ui.app.component.TextFieldDropdown
 import com.alexvanyo.composelife.ui.app.resources.DarkTheme
 import com.alexvanyo.composelife.ui.app.resources.DarkThemeConfig
 import com.alexvanyo.composelife.ui.app.resources.FollowSystem
 import com.alexvanyo.composelife.ui.app.resources.LightTheme
 import com.alexvanyo.composelife.ui.app.resources.Strings
+import com.alexvanyo.composelife.ui.mobile.component.DropdownOption
+import com.alexvanyo.composelife.ui.mobile.component.TextFieldDropdown
 import com.livefront.sealedenum.GenSealedEnum
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.launch
