+++ ui-mobile/src/jbMain/kotlin/com/alexvanyo/composelife/ui/mobile/component/TextField.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alexvanyo.composelife.ui.app.component
+package com.alexvanyo.composelife.ui.mobile.component
 
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.interaction.InteractionSource
