+++ ui-mobile/src/jvmMain/kotlin/com/alexvanyo/composelife/ui/mobile/component/DropdownOption.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alexvanyo.composelife.ui.app.component
+package com.alexvanyo.composelife.ui.mobile.component
 
 import androidx.compose.runtime.Composable
 import com.alexvanyo.composelife.parameterizedstring.ParameterizedString
